third_party/blink/web_tests/platform/win/external/wpt/clipboard-apis/clipboard-copy-selection-line-break.https_1-10-expected.txt

This is a testharness.js-based test.
[FAIL] test preformatted text with line breaks: "\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\rSecond" but got "First\\r\\nSecond"
[FAIL] test preformatted text with line breaks: "\\r\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\rSecond" but got "First\\r\\n\\r\\nSecond"
[FAIL] test preformatted text with line breaks: "\\n\\r"
  assert_equals: Check result from DataTransfer in paste event expected "First\\r\\n\\rSecond" but got "First\\r\\n\\r\\nSecond"
Harness: the test ran to completion.
